[P4ADEV-4530] add CategoryUtils
