+++ src/main/java/it/gov/pagopa/pu/debtpositions/service/TaxonomyValidatorServiceImpl.java
@@ -2,15 +2,16 @@ package it.gov.pagopa.pu.debtpositions.service;
 
 import it.gov.pagopa.pu.debtpositions.connector.organization.service.OrganizationService;
 import it.gov.pagopa.pu.debtpositions.connector.organization.service.TaxonomyService;
+import it.gov.pagopa.pu.debtpositions.util.CategoryUtils;
 import it.gov.pagopa.pu.debtpositions.util.SecurityUtils;
-import it.gov.pagopa.pu.debtpositions.util.Utilities;
 import it.gov.pagopa.pu.organization.dto.generated.Organization;
 import it.gov.pagopa.pu.organization.dto.generated.PagedModelTaxonomy;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 import org.springframework.util.CollectionUtils;
 
-import static it.gov.pagopa.pu.debtpositions.util.Utilities.getTaxonomyCodeFromCategory;
+import static it.gov.pagopa.pu.debtpositions.util.CategoryUtils.getTaxonomyCodeFromCategory;
+
 
 @Service
 @Slf4j
@@ -37,7 +38,7 @@ public class TaxonomyValidatorServiceImpl implements TaxonomyValidatorService {
   }
 
   public boolean isTaxonomyCodeValid(String taxonomyCode, String orgTypeCode) {
-      if(Utilities.CATEGORY_ALLOWED_PREFIXES.stream().noneMatch(taxonomyCode::startsWith) || !taxonomyCode.endsWith(Utilities.CATEGORY_SUFFIX)) {
+      if(CategoryUtils.CATEGORY_ALLOWED_PREFIXES.stream().noneMatch(taxonomyCode::startsWith) || !taxonomyCode.endsWith(CategoryUtils.CATEGORY_SUFFIX)) {
         log.info("The taxonomy code [{}] does not meet the required format", taxonomyCode);
         return false;
       }
