+++ src/main/java/it/gov/pagopa/pu/debtpositions/util/InstallmentUtils.java
@@ -129,7 +129,7 @@ public class InstallmentUtils {
     DebtPositionOrigin.REPORTING_PAGOPA
   );
 
-  public static final List<DebtPositionOrigin> SPONTANEOUS_DEBT_POSITION_ORIGINS = Stream.concat(
+  public static final List<DebtPositionOrigin> ORDINARY_CITIZEN_DEBT_POSITION_ORIGINS = Stream.concat(
     ORDINARY_CITIZEN_DEBT_POSITION_ORIGINS_NO_MIXED.stream(),
     Stream.of(
       DebtPositionOrigin.SPONTANEOUS_MIXED)
